Implementaion of printf
Using macro in "stdarg.h", which declares a "type " and define
macros for unknown amounts of arguments given to the called
function.

For example: printf("%d %d, a, b);
Theres two parameters which is given to "printf", so we should use
va_list to store these parameters. Reference to C99 standards.